+++ CHANGELOG.md
@@ -3,3 +3,5 @@ v3.0.0-alpha.1
 
 * Added GitHub workflow scaffolding. [#3](https://github.com/kalabox/lando/issues/3)
 * Added CLI build/package task. [#6](https://github.com/kalabox/lando/issues/6)
+* Added installer build tasks. [#5](https://github.com/kalabox/lando/issues/5)
+* Added testing framework and tasks. [#5](https://github.com/kalabox/lando/issues/5)